+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Fixed
+
+- Fix an issue with `beetcamp` CLI where the default config was missing.
+
 ## [0.24.0] 2026-03-03
 
 ### Removed